Deduplicate some networking code
