+++ components/core/src/clp_s/search/OutputHandler.cpp
@@ -4,6 +4,7 @@
 
 #include <spdlog/spdlog.h>
 
+#include "../../clp/networking/socket_utils.hpp"
 #include "../../reducer/CountOperator.hpp"
 #include "../../reducer/Record.hpp"
 #include "../../reducer/ReducerNetworkUtils.hpp"
@@ -15,38 +16,7 @@ NetworkOutputHandler::NetworkOutputHandler(
         bool should_output_timestamp
 )
         : OutputHandler(should_output_timestamp, true) {
-    struct addrinfo hints = {};
-    // Address can be IPv4 or IPV6
-    hints.ai_family = AF_UNSPEC;
-    // TCP socket
-    hints.ai_socktype = SOCK_STREAM;
-    hints.ai_flags = 0;
-    hints.ai_protocol = 0;
-    struct addrinfo* addresses_head = nullptr;
-    int error = getaddrinfo(host.c_str(), port.c_str(), &hints, &addresses_head);
-    if (0 != error) {
-        SPDLOG_ERROR("Failed to get address information for the server, error={}", error);
-        throw OperationFailed(ErrorCode::ErrorCodeFailureNetwork, __FILE__, __LINE__);
-    }
-
-    // Try each address until a socket can be created and connected to
-    for (auto curr = addresses_head; nullptr != curr; curr = curr->ai_next) {
-        // Create socket
-        m_socket_fd = socket(curr->ai_family, curr->ai_socktype, curr->ai_protocol);
-        if (-1 == m_socket_fd) {
-            continue;
-        }
-
-        // Connect to address
-        if (connect(m_socket_fd, curr->ai_addr, curr->ai_addrlen) != -1) {
-            break;
-        }
-
-        // Failed to connect, so close socket
-        close(m_socket_fd);
-        m_socket_fd = -1;
-    }
-    freeaddrinfo(addresses_head);
+    m_socket_fd = clp::networking::connect_to_server(host, port);
     if (-1 == m_socket_fd) {
         SPDLOG_ERROR("Failed to connect to the server, errno={}", errno);
         throw OperationFailed(ErrorCode::ErrorCodeFailureNetwork, __FILE__, __LINE__);
